Org babel tangle: fix `index/rebuild-emacs-init`
Added a few parameters to the function's `org-babel-tangle-file` call to
ensure it will compile to the right file, since it now put it into the
repository's directory because I follow the symlinks.

For that I needed to remove all `:tangle yes` headers from the
SRC-blocks.  Every SRC-block that did not have a `:tangle` header, I
changed them to be EXAMPLE-blocks.

This now compiles like a charm every time to the correct file that I
specified.