Functions: align column and join line improvements
index/align-column-dwim now works works as I mean it.  It operates on
the first matched candidate and not every candidate.  The latter can be
useful-- as I had it the default.  Mostly it annoyed me.

index/join-line supports operating on the region.  If you have a region
selected, every line will be joined together.