Byte compilation: now byte compiles
The kill-emacs-hook-- called when Emacs is terminated-- will now compile
a `config.elc` file instead of a `config.el` file.  This has the added
bonus that byte compiled lisp code is faster and in theory should make
all my customization faster.

This does add maybe triple the amount of time to execute the function.
But since this will get called before closing Emacs, it's fine.

The last commit fa30a24aae786ceb67aa97f057d2e7d7ac5b7708 added in the
`init.el` the `config.elc` already; so this is complimentary to it.