Text: update abstract text; add SRC-Block counter
Updated the abstract text at the top of the document.  Added a small
counter on the top that counts the amount of SRC-Blocks there are in
this document.  Removed both `init.el` and `early-init.el` contents from
the document literally and now link to the file.

There might be a way to let it be automatically inserted when exported
like how the inline SRC-Block works.  Sadly this can't be used as it
only handles one line output.