Text: minor changes
Added a bit more verbosity around the essentials section.  Deprecated a
few things; for a closer look, look at the diff.

Added two Org Mode settings: now the markup spans up to five lines and
not the default of two lines.  This might get increased in the future
when I need it.