Org: jump to section header function
This function allows me to jump to any header in the current org mode
document.  This can be nice when you know you want to go to "Elfeed" for
example.  Simply hit C-c C-j and it will ask for completion.  If you
want to go back you can simply pop the mark (C-u C-SPC).