Startup: add safety measures
When Emacs will recompile the file config.org to the target
config.elc, it will now check if the file is build; if not don't delete
the config.el file.

This will act as a fallback to the config.elc when it is not build,
since the compiling is more strict and when there is an error it won't
produce the desired file.  This is good in interactive use, because then
you know stuff that you wouldn't know before such as unused variables.
In an automatic manner this is bad when it doesn't produce a file.

Because of that I added these safety measures.

Now Emacs tries to load three different files: config.elc, config.el,
and config.org.  If none are found throw an error.