Magit: add magit-todos; add :commands
Magit Todos allows you to see all the possible todos in magit-status.
This will appear as another section that you can toggle.

There was an issue with magit-edit-line-commit, as it asked if I want to
enable the disabled command.  This was weird since I already enabled
that in the leaf block.  Now it loads the configuration on that command
too.