Mu4e: Org send mail
These days most emails are written in HTML.  This has the nice benefit
that you can use CSS to style your emails; the main reason is that it
supports every Org Mode markup: bold, hyperlinks, tables, and even
inline images.  Every nifty thing Org can do now you can send it as a
email.

I'm still in favor of text-only mail as it is very simplistic and the
reader should focus more on the text, then the markup.