Selectrum: mini-frame
mini-frame is a package that allows repositioning of the mini buffer
windows in a frame.  This allows Selectrum to be placed everywhere.
When I used ivy back in the past I loved the ivy-posframe package, that
allowed me to replace the ivy window; this is somewhat the same.

Why?

If I want to do something in the minibuffer, my eyes had to move at the
echo area.  This converted my attention from the task at point to the
minibuffer.  When I was done I had to move my eyes back to the place
where I was before-- this was rather troublesome and annoying.  Now it
plops right in the middle where I can easily switch my attention to.