Elisp: mapcar => --map; mapconcat => s-join
s-join is a little bit slower than mapconcat, but in my opinion it looks
way more elegant and cleaner; run the test yourself:

  (abs (- (car (benchmark-run 10000
                 (mapconcat #'identity '("a" "b" "c") " ")))
          (car (benchmark-run 10000
                 (s-join " " '("a" "b" "c"))))))

The result aka the difference between them is “0.003931496” over 10000
function calls-- basically nothing.

mapcar to dash's --map as it is, again, more elegant than the former.
Specifically that Emacs lisp currently does not support short hand
lambdas: (lambda (%) (quote %)) => #(quote %); like Clojure does.

With --map: % => it.  This can't be used with multiple arguments, but
mostly that is not what is needed.