Recompile: move saving the SHA into the unless
I didn't do that at first, because I thought when `s-equals?` tries to
compare a string to nil it would throw an error.  But it doesn't and
will simply return nil thus invoking the `unless` block.

Now the document will be rebuild when the `config.org.sha` file doesn't
exist.

This now doesn't write to disk when the SHA didn't change.