Modeline: do not print encoding info
This seems kinda redundant, you could make it smarter, and print the
encoding only when it isn't UTF-8, but I'm way to lazy.  Though I
remember it was in someones configuration, but that is a wild guess.
I'm leaving the information to it nonetheless, as the configuration is
quite good.

URL: https://github.com/tecosaur/emacs-config/