God Mode: initial settings
Trying out God Mode again.  Mapping it to M-SPC seems quite good, as the
standard function of that is to remove all spaces and add one,
effectively shrinking it to one.  Though this is nice with the C--,
`negative-argument', as it then spans over new lines too, removing
multiple new line in the process.